Extract e2e to its own file
